refactor: update import paths for Issues and PointsCalculator services for consistency
